Liouville vectors marginal functions

Marginal density, distribution, survival and inverse survival functions for Liouville copulas or Liouville vectors. The inverse survival function of Liouville vectors is not available in closed-form and is obtained numerically by root-finding. As such, Monte-Carlo approximation have been considered for dealing with inference to avoid computational bottlenecks. Note: the arguments of sliouv are reversed since they are meant to be called inside optim. The functions borrow \(psi\) functions and their derivatives from the copula-package.

sliouvm(x, family, alpha, theta)

pliouvm(x, family, alpha, theta)

isliouvm(u, family, alpha, theta)

dliouvm(x, family, alpha, theta)

Arguments

x
vector of quantiles from a Liouville copula (or a Liouville vector for the survival function , with support on the positive real line)

family
family of the Liouville copula. Either "clayton", "gumbel", "frank", "AMH" or "joe"

alpha
integer Dirichlet parameter

theta
parameter of the corresponding Archimedean copula

u
vector of quantiles or survival probabilities, (pseudo)-uniform variates

Value

a vector with the corresponding quantile, probability, survival probabilities

Examples

# NOT RUN {
#Marginal density
samp <- rliouv(n = 100, family = "clayton", alphavec <- c(2,3), theta = 2)
dliouvm(x=samp[,1], family="clayton", alpha=2, theta=2)
sum(log(dliouvm(x=samp[,1], family="clayton", alpha=2, theta=2)))
#Marginal distribution and (inverse) survival function
x <- rliouv(n = 100, family = "gumbel", alphavec <- c(2,3), theta = 2)
pliouvm(x[,1], family="gumbel", alpha=alphavec[1], theta=2)
su <- sliouvm(1-x[,1], family="gumbel", alpha=alphavec[1], theta=2)
isliouvm(u=su, family="clayton", alpha=2, theta=2)
#pliouv is the same as sliouv(isliouvm)
# }