Spectral density of the CDA of survival copula and copula of Liouville vectors

Computes the Liouville EV model or the scaled Dirichlet EV model spectral density

hmvevdliouv(w, alpha, rho, CDA = c("C", "S"), logdensity = FALSE)

Arguments

w
matrix of points in the unit simplex at which to evaluate the density

alpha
vector of Dirichlet allocations (strictly positive).

rho
parameter of limiting model corresponding to index of regular variation

CDA
copula domain of attraction of either the Liouville copula, C, or its survival copula S

logdensity
logical; whether to return the log density or not

Value

a vector with the same number of rows as w.

Examples

hmvevdliouv(seq(0.01,0.99,by=0.01), alpha=c(1,2), rho=0.2, CDA="C")

#>  [1] 3.744687e-15 1.071904e-12 3.075272e-11 3.442695e-10 2.302510e-09
#>  [6] 1.112250e-08 4.294035e-08 1.407466e-07 4.072417e-07 1.068248e-06
#> [11] 2.589089e-06 5.879877e-06 1.264624e-05 2.597286e-05 5.127310e-05
#> [16] 9.780482e-05 1.810498e-04 3.263926e-04 5.747298e-04 9.909097e-04
#> [21] 1.676268e-03 2.787002e-03 4.560736e-03 7.354352e-03 1.169694e-02
#> [26] 1.836246e-02 2.846707e-02 4.359547e-02 6.595868e-02 9.858014e-02
#> [31] 1.454979e-01 2.119540e-01 3.045206e-01 4.310803e-01 6.005522e-01
#> [36] 8.222348e-01 1.104653e+00 1.453872e+00 1.871393e+00 2.351963e+00
#> [41] 2.881860e+00 3.438346e+00 3.990865e+00 4.504180e+00 4.943018e+00
#> [46] 5.277220e+00 5.486133e+00 5.561147e+00 5.505916e+00 5.334465e+00
#> [51] 5.067952e+00 4.731009e+00 4.348459e+00 3.942872e+00 3.533117e+00
#> [56] 3.133783e+00 2.755270e+00 2.404264e+00 2.084419e+00 1.797060e+00
#> [61] 1.541839e+00 1.317284e+00 1.121234e+00 9.511643e-01 8.044181e-01
#> [66] 6.783662e-01 5.705055e-01 4.785172e-01 4.002957e-01 3.339569e-01
#> [71] 2.778337e-01 2.304637e-01 1.905731e-01 1.570582e-01 1.289663e-01
#> [76] 1.054785e-01 8.589235e-02 6.960682e-02 5.610855e-02 4.495970e-02
#> [81] 3.578740e-02 2.827442e-02 2.215119e-02 1.718885e-02 1.319333e-02
#> [86] 1.000024e-02 7.470459e-03 5.486449e-03 3.949017e-03 2.774602e-03
#> [91] 1.892944e-03 1.245098e-03 7.817320e-04 4.616881e-04 2.507441e-04
#> [96] 1.205633e-04 4.779647e-05 1.331668e-05 1.566215e-06

hmvevdliouv(seq(0.01,0.99,by=0.01), alpha=c(0.1,2), rho=0.2, CDA="S")

#>  [1] 1.524399e+00 1.105621e+00 9.261825e-01 8.231490e-01 7.557750e-01
#>  [6] 7.084207e-01 6.736435e-01 6.473989e-01 6.272805e-01 6.117588e-01
#> [11] 5.998119e-01 5.907301e-01 5.840050e-01 5.792635e-01 5.762265e-01
#> [16] 5.746823e-01 5.744684e-01 5.754594e-01 5.775583e-01 5.806905e-01
#> [21] 5.847990e-01 5.898415e-01 5.957874e-01 6.026164e-01 6.103171e-01
#> [26] 6.188857e-01 6.283251e-01 6.386449e-01 6.498605e-01 6.619931e-01
#> [31] 6.750692e-01 6.891212e-01 7.041865e-01 7.203086e-01 7.375364e-01
#> [36] 7.559252e-01 7.755363e-01 7.964379e-01 8.187053e-01 8.424213e-01
#> [41] 8.676770e-01 8.945721e-01 9.232156e-01 9.537266e-01 9.862347e-01
#> [46] 1.020881e+00 1.057819e+00 1.097213e+00 1.139241e+00 1.184096e+00
#> [51] 1.231980e+00 1.283109e+00 1.337707e+00 1.396006e+00 1.458238e+00
#> [56] 1.524630e+00 1.595388e+00 1.670687e+00 1.750641e+00 1.835271e+00
#> [61] 1.924463e+00 2.017896e+00 2.114963e+00 2.214652e+00 2.315399e+00
#> [66] 2.414902e+00 2.509896e+00 2.595913e+00 2.667049e+00 2.715819e+00
#> [71] 2.733221e+00 2.709193e+00 2.633681e+00 2.498492e+00 2.299886e+00
#> [76] 2.041422e+00 1.735958e+00 1.405404e+00 1.077248e+00 7.784216e-01
#> [81] 5.288109e-01 3.373281e-01 2.020690e-01 1.137564e-01 6.023108e-02
#> [86] 2.999649e-02 1.403376e-02 6.147934e-03 2.507339e-03 9.433104e-04
#> [91] 3.229723e-04 9.867231e-05 2.613798e-05 5.750782e-06 9.824091e-07
#> [96] 1.163262e-07 7.713903e-09 1.774246e-10 3.067865e-13

hmvevdliouv(seq(0.01,0.99,by=0.01), alpha=c(1,2), rho=0.2, CDA="S")

#>  [1] 2.371429e-08 4.073746e-07 2.215845e-06 7.530048e-06 1.978204e-05
#>  [6] 4.417392e-05 8.819960e-05 1.622926e-04 2.806283e-04 4.621143e-04
#> [11] 7.316124e-04 1.121441e-03 1.673222e-03 2.440145e-03 3.489743e-03
#> [16] 4.907286e-03 6.799926e-03 9.301755e-03 1.257996e-02 1.684229e-02
#> [21] 2.234616e-02 2.940955e-02 3.842424e-02 4.987162e-02 6.434150e-02
#> [26] 8.255447e-02 1.053880e-01 1.339067e-01 1.693968e-01 2.134045e-01
#> [31] 2.677773e-01 3.347064e-01 4.167667e-01 5.169493e-01 6.386765e-01
#> [36] 7.857887e-01 9.624832e-01 1.173181e+00 1.422292e+00 1.713837e+00
#> [41] 2.050902e+00 2.434868e+00 2.864441e+00 3.334487e+00 3.834809e+00
#> [46] 4.349062e+00 4.854163e+00 5.320598e+00 5.714101e+00 5.998935e+00
#> [51] 6.142705e+00 6.122003e+00 5.927644e+00 5.567963e+00 5.068862e+00
#> [56] 4.470219e+00 3.819393e+00 3.163591e+00 2.543097e+00 1.986929e+00
#> [61] 1.511453e+00 1.121560e+00 8.134058e-01 5.776570e-01 4.024184e-01
#> [66] 2.754358e-01 1.854793e-01 1.230265e-01 8.044922e-02 5.189806e-02
#> [71] 3.304234e-02 2.076660e-02 1.288321e-02 7.887549e-03 4.763462e-03
#> [76] 2.835807e-03 1.662722e-03 9.591051e-04 5.435289e-04 3.021159e-04
#> [81] 1.643861e-04 8.735370e-05 4.520783e-05 2.271005e-05 1.102958e-05
#> [86] 5.153853e-06 2.303339e-06 9.773104e-07 3.900414e-07 1.446731e-07
#> [91] 4.909187e-08 1.491747e-08 3.939445e-09 8.653260e-10 1.477072e-10
#> [96] 1.748401e-11 1.159276e-12 2.666327e-14 4.610352e-17