Women daily nutrient intake

The nutrient data frame consists of quintuples consisting of four day measurements for intake of calcium, iron, protein, vitamin A and C from women aged 25 to 50 in the United States as part of the "Continuing Survey of Food Intakes of Individuals" program. The processed data has 737 measurements from a cohort study of the United States Department of Agriculture (USDA) and it is available online from the University of Pennsylvania repository.

nutrient

Format

This data frame contains the following columns:

id

A numeric vector containing the identification of the participant.

calcium

A numeric vector containing measured calcium.

iron

A numeric vector containing iron measurements.

protein

A numeric vector containing protein measurements.

vitamin.a

A numeric vector containing vitamin A measurements.

vitamin.c

A numeric vector containing vitamin C measurements.

Source

The survey data was processed by Dr. Andrew Wiesner for a course on multivariate statistics at The Pennsylvania State University.

References

Genest, C., Neslehova, J., Quessy, J.-F. (2012) Tests of symmetry for bivariate copulas. Annals of the Institute of Statistical Mathematics, 64(4), 811--834.

Genest, C. and Neslehova, J. G. (2013) Assessing and Modeling Asymmetry in Bivariate Continuous Data. In P. Jaworski, F. Durante, & W. K. Hardle (Eds.), Copulae in Mathematical and Quantitative Finance, Lecture Notes in Statistics, 91--114, Springer: Berlin Heidelberg.

Li, B., Genton, M. G. (2013) Nonparametric identification of copula structures. Journal of the American Statistical Association, 108(502), 666--675.