Pickands dependence function for the copula domain of attraction of Liouville survival copulas

Pickands dependence function as in Belzile (2014), Proposition 40 and Example 4 and Belzile (2014), Proposition 41, assuming that the parameter alpha is integer-valued. Returns the Pickands dependence function of the copula domain of attraction (CDA) of the survival copula, the scaled Dirichlet extreme value model, or the CDA of the copula, the Liouville EV model.

pickands.liouv(t, rho = 0.5, alpha = c(1, 1), CDA = c("C", "S"))

Arguments

t
pseudo-angle in (0,1)

rho
index of regular variation parameter

alpha
vector of Dirichlet allocations. Currently must be of length 2

CDA
select the extremal attractor of the copula (C) or the survival copula (S)

Value

value of Pickands function for the scaled Dirichlet EV model

Examples

pickands.liouv(seq(0,1,by=0.01),1,c(0.1,0.3),CDA="S")

#>   [1] 1.0000000 0.9927850 0.9868409 0.9815611 0.9767653 0.9723564 0.9682723
#>   [8] 0.9644694 0.9609148 0.9575831 0.9544535 0.9515093 0.9487361 0.9461217
#>  [15] 0.9436558 0.9413291 0.9391336 0.9370621 0.9351084 0.9332666 0.9315316
#>  [22] 0.9298987 0.9283636 0.9269224 0.9255716 0.9243079 0.9231281 0.9220296
#>  [29] 0.9210098 0.9200661 0.9191964 0.9183986 0.9176708 0.9170110 0.9164178
#>  [36] 0.9158894 0.9154243 0.9150213 0.9146789 0.9143960 0.9141714 0.9140041
#>  [43] 0.9138930 0.9138373 0.9138360 0.9138883 0.9139935 0.9141508 0.9143597
#>  [50] 0.9146194 0.9149294 0.9152892 0.9156984 0.9161564 0.9166629 0.9172176
#>  [57] 0.9178201 0.9184701 0.9191675 0.9199121 0.9207036 0.9215420 0.9224273
#>  [64] 0.9233593 0.9243381 0.9253638 0.9264365 0.9275563 0.9287234 0.9299382
#>  [71] 0.9312008 0.9325117 0.9338713 0.9352802 0.9367388 0.9382479 0.9398081
#>  [78] 0.9414204 0.9430857 0.9448050 0.9465796 0.9484109 0.9503003 0.9522496
#>  [85] 0.9542607 0.9563360 0.9584780 0.9606897 0.9629746 0.9653366 0.9677806
#>  [92] 0.9703125 0.9729392 0.9756696 0.9785150 0.9814903 0.9846165 0.9879245
#>  [99] 0.9914657 0.9953447 1.0000000

pickands.liouv(t = seq(0,1,by=0.01), rho = 0.5, alpha = c(1,3), CDA="C")

#>   [1] 1.0000000 0.9900000 0.9800000 0.9700000 0.9600001 0.9500004 0.9400014
#>   [8] 0.9300035 0.9200081 0.9100169 0.9000325 0.8900587 0.8801006 0.8701649
#>  [15] 0.8602597 0.8503954 0.8405841 0.8308401 0.8211794 0.8116201 0.8021818
#>  [22] 0.7928857 0.7837538 0.7748092 0.7660752 0.7575752 0.7493322 0.7413686
#>  [29] 0.7337056 0.7263633 0.7193601 0.7127128 0.7064364 0.7005437 0.6950457
#>  [36] 0.6899515 0.6852681 0.6810006 0.6771524 0.6737251 0.6707188 0.6681321
#>  [43] 0.6659624 0.6642058 0.6628574 0.6619115 0.6613613 0.6611997 0.6614189
#>  [50] 0.6620105 0.6629660 0.6642763 0.6659324 0.6679250 0.6702448 0.6728823
#>  [57] 0.6758284 0.6790736 0.6826088 0.6864250 0.6905132 0.6948647 0.6994710
#>  [64] 0.7043236 0.7094145 0.7147358 0.7202796 0.7260387 0.7320056 0.7381734
#>  [71] 0.7445353 0.7510847 0.7578152 0.7647208 0.7717955 0.7790335 0.7864295
#>  [78] 0.7939780 0.8016740 0.8095124 0.8174887 0.8255982 0.8338364 0.8421992
#>  [85] 0.8506825 0.8592824 0.8679950 0.8768167 0.8857441 0.8947737 0.9039023
#>  [92] 0.9131268 0.9224441 0.9318515 0.9413460 0.9509250 0.9605859 0.9703262
#>  [99] 0.9801435 0.9900355 1.0000000