Multiple marginal survival function for Liouville vectors

This function is a wrapper around sliouv; it allows the user to treat all the data matrix simultaneously by applying different parameters to each margin.

sliouv_m(x, family, alphavec, theta)

Arguments

x
sample from copula

family
family of the Liouville copula. Either "clayton", "gumbel", "frank", "AMH" or "joe"

alphavec
vector of Dirichlet allocations (must be a vector of integers)

theta
parameter of the corresponding Archimedean copula

Value

a matrix of same length as x with the survival probabilities

Examples

x <- rliouv(n = 100, family = "gumbel", alphavec <- c(2,3), theta = 2)
sliouv_m(x, family="gumbel", alphavec=c(2,3), theta=2)

#>             [,1]      [,2]
#>   [1,] 0.6250287 0.8345174
#>   [2,] 0.8582687 0.9009348
#>   [3,] 0.9408536 0.9206452
#>   [4,] 0.7339179 0.7925883
#>   [5,] 0.7734604 0.8168547
#>   [6,] 0.7327522 0.7439268
#>   [7,] 0.8793051 0.9051811
#>   [8,] 0.7113770 0.7548505
#>   [9,] 0.6085502 0.6663007
#>  [10,] 0.8622927 0.9784987
#>  [11,] 0.6205253 0.7904744
#>  [12,] 0.5848836 0.6618892
#>  [13,] 0.7165320 0.8125770
#>  [14,] 0.9022668 0.8548283
#>  [15,] 0.5854140 0.6854638
#>  [16,] 0.6321860 0.8124262
#>  [17,] 0.6852768 0.7672225
#>  [18,] 0.7236696 0.8150003
#>  [19,] 0.7859941 0.8634752
#>  [20,] 0.5842836 0.7152930
#>  [21,] 0.6229467 0.7869486
#>  [22,] 0.6811788 0.7191081
#>  [23,] 0.6079458 0.6869632
#>  [24,] 0.6106181 0.7097737
#>  [25,] 0.7725835 0.7575117
#>  [26,] 0.5747181 0.6633152
#>  [27,] 0.7968928 0.9071963
#>  [28,] 0.8060413 0.8468020
#>  [29,] 0.5628840 0.6538097
#>  [30,] 0.9182162 0.9065740
#>  [31,] 0.6277241 0.6943496
#>  [32,] 0.6199468 0.6591963
#>  [33,] 0.7199941 0.8036774
#>  [34,] 0.7229278 0.7808853
#>  [35,] 0.7222907 0.7975484
#>  [36,] 0.6016086 0.7350814
#>  [37,] 0.5854551 0.6965276
#>  [38,] 0.6399677 0.7340434
#>  [39,] 0.8568692 0.9293818
#>  [40,] 0.7280595 0.7392982
#>  [41,] 0.7478314 0.7453871
#>  [42,] 0.7466964 0.8618211
#>  [43,] 0.9220651 0.8843441
#>  [44,] 0.9234527 0.9288285
#>  [45,] 0.7746505 0.8214229
#>  [46,] 0.6418720 0.8080244
#>  [47,] 0.8071646 0.8901588
#>  [48,] 0.7468291 0.8061725
#>  [49,] 0.7371777 0.7738444
#>  [50,] 0.7863615 0.7479045
#>  [51,] 0.6902770 0.7566930
#>  [52,] 0.6842299 0.8012749
#>  [53,] 0.7125212 0.7592644
#>  [54,] 0.5543705 0.6444759
#>  [55,] 0.8288858 0.8426101
#>  [56,] 0.5882620 0.6568239
#>  [57,] 0.6197132 0.7033623
#>  [58,] 0.6682942 0.7168309
#>  [59,] 0.6345073 0.6887164
#>  [60,] 0.8872252 0.9071751
#>  [61,] 0.5555778 0.6679329
#>  [62,] 0.7343083 0.7580395
#>  [63,] 0.8104929 0.8818425
#>  [64,] 0.7090282 0.8664624
#>  [65,] 0.7214541 0.7340923
#>  [66,] 0.6882490 0.8615860
#>  [67,] 0.5638890 0.6631742
#>  [68,] 0.7197482 0.7650895
#>  [69,] 0.5693568 0.6564733
#>  [70,] 0.8789581 0.8556771
#>  [71,] 0.9845782 0.9452069
#>  [72,] 0.6603744 0.6762928
#>  [73,] 0.6246033 0.6932609
#>  [74,] 0.6514400 0.7275403
#>  [75,] 0.8716832 0.8102194
#>  [76,] 0.7636679 0.8796089
#>  [77,] 0.6101042 0.6715370
#>  [78,] 0.7230251 0.8030928
#>  [79,] 0.7183811 0.7571112
#>  [80,] 0.9721307 0.8531075
#>  [81,] 0.5663092 0.6808391
#>  [82,] 0.6264022 0.6794883
#>  [83,] 0.6019439 0.6746145
#>  [84,] 0.5587627 0.6478601
#>  [85,] 0.5558302 0.6472352
#>  [86,] 0.6796875 0.8529120
#>  [87,] 0.6837043 0.7199623
#>  [88,] 0.5793588 0.6701022
#>  [89,] 0.6193462 0.6980755
#>  [90,] 0.6394877 0.7700769
#>  [91,] 0.8094286 0.8496850
#>  [92,] 0.6171678 0.7528215
#>  [93,] 0.6408703 0.7370564
#>  [94,] 0.6638070 0.6972857
#>  [95,] 0.6830614 0.7021625
#>  [96,] 0.6555081 0.7503131
#>  [97,] 0.7027355 0.9298399
#>  [98,] 0.8538953 0.8631215
#>  [99,] 0.7103029 0.7904778
#> [100,] 0.5652285 0.6556836

all(sliouv_m(x, family="gumbel", alphavec=c(2,3), theta=2)[,1]-
  sliouvm(x[,1], family="gumbel", alpha=2, theta=2)==0)

#> [1] TRUE