Parametric bootstrap confidence interval for the parameter theta for Liouville copula

The parametric bootstrap provides confidence intervals by repeatedly sampling datasets from the postulated Liouvilla copula model. If \(d=2\) and the model is either gumbel or clayton, the value of Kendall's \(\tau\) is calculated from the sample, and the confidence interval or the quantiles correspond to the inverse \(\tau^{-1}(\tau(\theta))\) for the bootstrap quantile values of \(\tau\) (using monotonicity).

theta.bci(B = 1999, family, alphavec, n, theta.hat, quant = c(0.025,
  0.975), silent = FALSE)

Arguments

B
number of bootstrap replicates

family
family of the Liouville copula. Either "clayton", "gumbel", "frank", "AMH" or "joe"

alphavec
vector of Dirichlet allocations (must be a vector of integers)

n
sample size

theta.hat
estimate of theta

quant
if the vector of probability is specified, the function will return the corresponding bootstrap quantiles

silent
boolean for output progress. Default is FALSE, which means iterations are printed if \(d>2\).

Value

a list with a 95 and the bootstrap values of Kendall's tau in boot_tau if \(d=2\) and the model is either gumbel or clayton. Otherwise, the list contains boot_theta.

Details

Since no closed-form formulas exist for the other models or in higher dimension, the method is extremely slow since it relies on maximization of a new sample from the model and look up the corresponding parameters.

Examples

# NOT RUN {
theta.bci(B=99, family="gumbel", alphavec=c(2,3), n=100, theta.hat=2)
theta.bci(B=19, family="AMH", alphavec=c(1,2), n=100, theta.hat=0.5, quant=c(0.05,0.95))
theta.bci(B=19, family="frank", alphavec=c(1,2,3), n=100, theta.hat=0.5, quant=c(0.05,0.95))
# }